Updated Manager Data and moved Modals into modals folder
