+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules
 .env
 employeeData.json
-managerData.json
 avatars
