+++ client/EntryPoint.jsx
@@ -5,17 +5,15 @@ import { Container, CssBaseline } from '@mui/material'
 import PageHeader from './components/PageHeader'
 import ManagerGrid from './components/ManagerGrid.jsx'
 import EmployeeGrid from './components/EmployeeGrid.jsx'
-import ManagerDetailsModal from './components/ManagerDetailsModal.jsx'
-import EmployeeDetailsModal from './components/EmployeeDetailsModal.jsx'
-import WeekendShiftsModal from './components/WeekendShiftsModal.jsx'
-import FinalShiftModal from './components/FinalShiftModal.jsx'
-import InputScheduleModal from './components/InputScheduleModal.jsx'
+import EmployeeDetailsModal from './components/modals/EmployeeDetailsModal.jsx'
+import WeekendShiftsModal from './components/modals/WeekendShiftsModal.jsx'
+import FinalShiftModal from './components/modals/FinalShiftModal.jsx'
+import InputScheduleModal from './components/modals/InputScheduleModal.jsx'
 
 export default function EntryPoint (props) {
   const previewToggle = false
 
   const [employeeId, setEmployeeId] = React.useState('')
-  const [managerId, setManagerId] = React.useState('')
   const [showWeekendModal, setShowWeekendModal] = React.useState(false)
   const [showFinalModal, setShowFinalModal] = React.useState(false)
   const [showInputModal, setShowInputModal] = React.useState(false)
@@ -23,12 +21,6 @@ export default function EntryPoint (props) {
   const onEmployeeClicked = (newEmployeeId) => {
     if (newEmployeeId !== employeeId) {
       setEmployeeId(newEmployeeId)
-    }
-  }
-
-  const onManagerClicked = (newManagerId) => {
-    if (newManagerId !== managerId) {
-      setManagerId(newManagerId)
     }
   }
 
@@ -46,7 +38,6 @@ export default function EntryPoint (props) {
 
   const onModalClose = () => {
     setEmployeeId('')
-    setManagerId('')
     setShowWeekendModal(false)
     setShowFinalModal(false)
     setShowInputModal(false)
@@ -56,9 +47,8 @@ export default function EntryPoint (props) {
     <Container>
       <CssBaseline />
       <PageHeader previewToggle={previewToggle} openWeekendModal={openWeekendModal} openFinalModal={openFinalModal} openInputModal={openInputModal} />
-      <ManagerGrid previewToggle={previewToggle} onClickCallback={onManagerClicked} />
+      <ManagerGrid previewToggle={previewToggle} onClickCallback={onEmployeeClicked} />
       <EmployeeGrid previewToggle={previewToggle} onClickCallback={onEmployeeClicked} />
-      <ManagerDetailsModal previewToggle={previewToggle} employeeId={managerId} onClose={onModalClose} />
       <EmployeeDetailsModal previewToggle={previewToggle} employeeId={employeeId} onClose={onModalClose} />
       <WeekendShiftsModal previewToggle={previewToggle} openModal={showWeekendModal} onClose={onModalClose} />
       <FinalShiftModal previewToggle={previewToggle} openModal={showFinalModal} onClose={onModalClose} />
