+++ server/api/dataCollection.js
@@ -10,9 +10,8 @@ dotenv.config()
 // Generates data of the current employees in humanity
 export function generateEmployees () {
   return new Promise((resolve, reject) => {
-    // Creates initial list of employees and managers
+    // Creates initial list of employees
     const employees = []
-    const managers = []
 
     // Axios call to the humanity api
     axios.get(`https://www.humanity.com/api/v2/employees?access_token=${process.env.API_KEY}`)
@@ -20,36 +19,17 @@ export function generateEmployees () {
         // Loops through response data
         response.data.data.forEach(employee => {
           // Checks if the employee does not have the manager label
-          if (!employee.schedules['1949304']) {
+          if (!employee.schedules['1949304'] || [''].includes(employee.name)) {
             // Creates image name based of the employee's name
             const imageName = employee.name.replace(/[^a-zA-Z]/ig, '')
             // Fetches image and places it in the public avatars folder
-            fetch(employee.avatar.medium)
-              .then(image => image.arrayBuffer())
-              .then(image => fs.writeFileSync(path.join('public', 'avatars', `${imageName}.jpg`), Buffer.from(image)))
+            if (employee.avatar.medium !== '/app/layout/images/no_avatar.png') {
+              fetch(employee.avatar.medium)
+                .then(image => image.arrayBuffer())
+                .then(image => fs.writeFileSync(path.join('public', 'avatars', `${imageName}.jpg`), Buffer.from(image)))
+            }
             // Creates an employee with data provided and pushes that employee to the employees array
             employees.push({
-              id: employee.id,
-              name: employee.name,
-              email: employee.email,
-              phone: employee.cell_phone,
-              locations: employee.schedules,
-              interestingFact: employee.custom['3678'].value,
-              yearAtSTAR: employee.custom['6879'].value,
-              major: employee.custom['6880'].value,
-              expectedGrad: employee.custom['15936'].value,
-              avatar: `${imageName}.jpg`
-            })
-          // Checks if the employee with the manager title has a name in the list provided (Names must be an exact match)
-          } else if ([''].includes(employee.name)) {
-            // Creates image name based of the employee's name
-            const imageName = employee.name.replace(/[^a-zA-Z]/ig, '')
-            // Fetches image and places it in the public avatars folder
-            fetch(employee.avatar.medium)
-              .then(image => image.arrayBuffer())
-              .then(image => fs.writeFileSync(path.join('public', 'avatars', `${imageName}.jpg`), Buffer.from(image)))
-            // Creates an manager with data provided and pushes that manager to the managers array
-            managers.push({
               id: employee.id,
               name: employee.name,
               email: employee.email,
@@ -65,7 +45,6 @@ export function generateEmployees () {
         })
         // Writes the managers and employees arrays to their appropriate file
         fs.writeFileSync(path.join('server', 'api', 'employeeData.json'), JSON.stringify(employees, null, 2))
-        fs.writeFileSync(path.join('server', 'api', 'managerData.json'), JSON.stringify(managers, null, 2))
         return resolve()
       })
       .catch((err) => {
